feat ( react-practice: Note-taker ) -- finished video 7
Issues caused by an update in react caused this tutorial to not function correctly, rather than just updating to the a specific version of ReactFire, as instructed below the video, to fix the app, I researched a different way to accomplish the same thing since, in the video, the instructor suggested there were several ways to accomplish the same thing.

I'm using push instead of set now.