Refactor to retrieve app data again when given username changes
finished video 11

I now know how to use three react lifecycle events: componentDidMount, componentWillUnmount, componentWillReceiveProps.

As I currently understand it:
componentDidMount happens when changing to a route that uses this component

componentWillUnmount happens when leaving a route that uses this component to a route that does not.

componentWillReceiveProps happens when the state information, namely the route, changes.