Refactor ES5 patterns to ES2015 patterns
Use import instead of require.

Use destructuring to import the Router and 2 properties of the Router from react-router.

== NOTES ==

Destructuring at first glance isn't particularly intuitive. Defining a "Router" var then referencing properties from it is more clear.

I am assuming the order of the keys in the destructuring pattern determines the parent from which each subsequent key will seek properties.

I.E.
1) Router = react-router object

2) Key "DefaultRoute" is sought on Router because it is the first key to it's left, if a "DefaultRoute" property is found on Router, it takes it as it's definition.

3) Key "Route" is sought on the key before it, "DefaultRoute"
Since DefaultRoute.Route is undefined ( I checked ;-D ), it goes on to try the next key to to the left Router.Route & if a "Route" property is found then it takes that as it's definition.

I would guess that, when all keys are exhausted, the destructured keys would NOT fallback to prototypical inheritance in their search for a value, but would instead return undefined or throw an error.

I still need to run more tests to and look at the destructuring documentation to determine the accuracy of my assertions.

MOST INTERESTING thing I noticed is that with "var Router..." I can console.log "Router" and see it's properties & methods, whereas with import I cannot log Router but I CAN log the value of the subsequent keys. Interesting!

I thought that this was might be an oversight of the tutorial author.

Because I wasn't using var Router, except to get the child properties from react-router, I removed the "Router" key from the destructure & saw no negative effects, the app still ran perfectly.

Because there is no "Router" property on the 'react-router' object it is undefined.

CONCLUSION

If you're planning on referencing any child properties of react-router you can't define them later, they have to be referenced by their actual names and not aliases in the destructuring pattern at import.

I could be wrong but this is the impression I'm getting at this stage in the tutorial! :-D